Share uploaded entries fixture with all tests
